Restore previously typed commit message when commit operation fails
Summary:
I hit an error in `sl commit` due to some weird network issue, and found my carefully typed commit message disappeared.

If we see the commit operation fail, let's restore your edited commit message back where it was. We can just yoink it from the Commit operation.

We probably need to do this several other places, actually:
commit, amend, metaedit, partial commit/amend

Let's start with commit.

Note that we need a way of looking for commands failing. We could try to do this in handlers for `runOperation`, but I kind of like the pattern of watching for operations to exit and having this as an "effect" there. So I added a way to subscribe to operation exits.

Reviewed By: quark-zju

Differential Revision: D54786161

fbshipit-source-id: d6c2723f9967ea511ae35dff0f75ed215e7d175f