Log metadata on the current configuration version [2/n]
Summary:
Decorate Scuba logs with the version of the repos config that was current in the context of processing that request. For long-lived requests that span a configuration push, that might be the previous version. Having this information handy is useful for troubleshooting.

This diff introduces a generic struct to hold config information, and logs it from generic code.
The actual content of the config depends on the specific server; this diff adds a new middleware for the LFS server, in the future we should do the same for Mononoke.

Differential Revision: D54537233

fbshipit-source-id: b17fdeed200bcf6bd12b97fb20fe2d27a5416272