Log metadata on the current configuration version [3/n]
Summary:
The previous diff in the stack (D54537233) adds the general infrastructure, but it doesn't actually compute a meaningful has of the content.

For now we will just serialize and take a hash of that. That will still allow us to e.g. track a version rollout through the different TW jobs / tasks, finding any stuck on an old version and so on.

It would be nice to be able to get hold of the original configuration on disk and hash it (or even better, ), but we don't have an API for that. T182154838 will track that as a followup.

Reviewed By: RajivTS

Differential Revision: D54743888

fbshipit-source-id: ed194e4da266110ba136ca306d70dfe15ffb4e88