Collapse use of ::types:: (E-H)
Summary:
This diff is part of the codemod for https://fb.workplace.com/groups/rust.language/posts/25780668961555042.

Thrift-generated data structures, clients, and service traits all used to be exported in a single crate, with the data structures being under a `types` module. Now that data structures have a crate of their own, the `types` module is vestigial: `foo::types` refers to the same contents as `foo`.

Generated using the code in P1195926494 and the procedure from D54838520.

Reviewed By: zertosh

Differential Revision: D54871741

fbshipit-source-id: 78a3cb7ea4283929b9664bcce14fdd76848a7765