CommitInfo: rename isHead to isDot
Summary:
The concept is named differently in different tools:

`man git`:
> SYMBOLIC IDENTIFIERS
> [...] HEAD: indicates the head of the current branch.

> DISCUSSION
> [...] A special ref named HEAD contains the name of the currently checked-out branch.

`hg help revisions`:
> The reserved name "." indicates the working directory parent.
> [...]
> "head()": Changeset is a named branch head. [note: sapling has no "named branch"]
> "heads(set)": Members of set with no children in set.

So in Git the "HEAD" is a branch name that resolves to a hash. In Mercurial /
Sapling, "head" is more about the concept of some commit that has no children.
".", "working (directory) parent", "dirstate p1".

Note: in the new dag concept, "You are here" is a virtual commit like "wdir()"
with "." being its parent. The "." commit is different from "You are here".

Migrate to Sapling/Mercurial termlogy to reduce confusion. I chose "dot"
instead of "working parent" or "you are here parent" since it's shorter and
seems unambiguous.

Reviewed By: evangrayk

Differential Revision: D54819891

fbshipit-source-id: 7a5b32b84f8c21f4f65cb38192d347cfd16595e1