drawdag: add ability to apply changes to all commits
Summary:
Currently drawdag action specs apply to specific commits, however it would be useful to allow specifications for default values that apply to all commits.

Allow this by accepting `*` as a commit node name that means "all commits" in action specifiers.  These are applied first, so that per-commit specifiers are able to override their values.

Differential Revision: D54729621

fbshipit-source-id: 67521cb9fcb7c6a6316070ee70f37562488570ea