Fix 'goto' after download not going to commit after rebasing
Summary:
After downloading, we may rebase. After rebasing, we try to goto. We need to goto the successor of the commit, in case you enter a hash instead of a diff number. This way, the goto will use the rebased commit, instead of the originally entered hash.

We only really need to use the successor when we're doing a rebase. Otherwise, might as well use the exact revset they typed.

Also, rename `enteredDiffNum` to `enteredRevset` since we support arbitrary revsets in here.

Reviewed By: quark-zju

Differential Revision: D54882067

fbshipit-source-id: 0257dede8766e8415822188deb338716329b3f6b