Generate x-repo-submodule metadata file
Summary:
This is the work for T179534458, which is about generating the "x-repo submodule metadata file" which will store the submodule commit being expanded in that revision.

While this wasn't needed in Phase 1 (i.e. just expanding), we thought it would be best to already implement it so all the commits with expansion have the metadata file.

## What do we store in the file?
Initially I just copied the content id from the `GitSubmodule` FileChange from the small repo. However, this hash is encoded there, which is not very user-friendly and will likely lead to issues when we want to examine the file in the editor or CodeHub (see [this submodule file in CodeHub](https://www.internalfb.com/code/whatsapp-iphone/submodules/voip?offset=1600&pagesize=100)).

## What's next
With this file being generated, I'll move on to T181473986, which is about reviewing and extending test coverage before we start using the submodule expansion.

Reviewed By: mitrandir77

Differential Revision: D54308547

fbshipit-source-id: 02176e073088fa0b7252583256e9af8981fe4a7a