Create module for public interface
Summary:
## This stack
This diff starts some refactorings to improve the x-repo library.

The library is very big and there are a lot of things that could be improved:

- Better interface definition, e.g. make it clear what is (and most importantly **what should be**) exported.
- Modularization and more structured dependency graph, e.g. have smaller modules instead of files with 2.5K lines where everything depends on everything.
  - The structure of modules and dependencies should represent abstractions.
- Reduce duplication (e.g. `CommitSyncer` and `CommitInMemorySyncer`)
- Organize code in general, e.g. sort public vs private methods, group similar methods.

All this will then make it easy to better understand the code and find more improvement opportunities, such as removing duplication and improving abstractions.

## This diff
I had a hard time deciding how I would start this stack, but I decided to go with setting one module to be used as a public interface for this library/crate.

In this diff I move the contents of the massive file `cross_repo_sync/src/lib.rs` to `cross_repo_sync/src/commit_syncers_lib.rs`. The former will be used as the public module, where we'll export only what we should.

I also update some internal imports to use this new module.

In the next diff I'll update the external imports to import only from the public interface.

IMPORTANT: I don't think that this diff and the next leave everything in a perfect state, but I do believe that it's **already better than the current state**. After D54856985 I'll split this massive `commit_syncers_lib.rs` file into multiple modules.

Reviewed By: mitrandir77

Differential Revision: D54855752

fbshipit-source-id: ff561c5767b9b48f30e1e26e082d32e302101a87