Test import of unroundtripable git commit
Summary:
Add an integration test that shows an issue with gitimport and data derivation with non-roundtrippable commits.

By unroundtripable, we mean: given a git commit, the bonsai generated from it doesn't contain enough information
to produce the exact same commit back when deriving the git commit from the bonsai.

We use the "encoding" field of git which is not present in bonsai to showcase this scenario.

Thanks to this test, we can see that gitimport fails to import non-roundtripable git commits.
The next diff description explains the situation in more details.
The next diff fixes the test.

Reviewed By: gustavoavena

Differential Revision: D54905695

fbshipit-source-id: 84b7f36f30086cb5f5954445a99060e77fcad230