Fix conversion of git time with timezone offset
Summary:
Mononoke stores the timezone offset in western convention, which is the opposite
of what we get from git, so we need this minus sign.

This is a single character change to fix the bug. The next diff refactors the code
so it's easier to reason about it. The next diff also adds a unit test that shows
we're not making 2 mistakes that cancel out, but indeed we do need this minus sign.

Reviewed By: RajivTS

Differential Revision: D54944772

fbshipit-source-id: 096dcb859501b9e1981ae585770d4786b4f8e158