fix implementation for fetching trees
Summary:
fix implementation for fetching trees

we return the content of manifest blob and the parents all the time even if
request doesn't require that.

So, fixing implementation to comply with the request's attributes correctly,
also matching correct implementation in the eager repo.

https://www.internalfb.com/code/fbsource/[2d2381cc4fafb0d0b178a6080bdfeea14b10466c]/fbcode/eden/scm/lib/eagerepo/src/api.rs?lines=244

the tree attributes consist of:

            manifest_blob
            parents
            child_metadata

before the fix the values of the manifest_blob flag and the parents flag were never checked and we just returned those parts always

Reviewed By: markbt

Differential Revision: D54948203

fbshipit-source-id: e99aff422a8a4323d2287dea91d80a66b90f9f57