alerts: show alerts when using errorredirect extension
Summary:
Display matching alerts when there is an errorredirect script configured.

This breaks errorredirect's assumption that it is handling the entire experience when there is a crash, but I like having "alerts" built in as a feature since we can control it precisely with config changes.

I also silenced the alerts output if HGPLAIN mode is activated. Doesn't seem like we want these messages showing up for automation.

Reviewed By: sggutier

Differential Revision: D54962995

fbshipit-source-id: a378e825f17c56ad4f3c9c50f50e893bd5d7ebe4