don't sync mapping change via extras for forward-synced repos
Summary:
The mapping change commit (empty commit with special extra) need to end up as a noop change when we try to sync it to the small repo when the small repo is still source of truth.

Otherwise we don't end up with clean remapping between large and small and that commit is a single dangling one on the side of the small repo commit graph.

I think the intention of the `if` check was already that. This commit fixes that logic.

Reviewed By: gustavoavena

Differential Revision: D54877760

fbshipit-source-id: 6a7945ba30280356bfbbf5aeee3ce6b5a5c9b7e1