validate working copy equivalence instead of assuming it
Summary:
This diff is basically defensive programming but the stakes are high enough
that I think is justified.

In the D54877757 I'm assuming that current target_bookmark commit is working
copy equivalent to the parent of the commit we're trying to sync. In this
commit I'm acutally checking the mapping to verify that fact.

Differential Revision: D54965081

fbshipit-source-id: 3c52fe33b86de9c4946b5ae3637d562fec9f41bf