Fix packfile stream generation inefficiencies
Summary:
When using the new Git server for clone or pull in a cache cold setting, the operation could take hours even for small repos due to transfer speed in the KBs. For large repos, even with caches warmed up the performance is suboptimal. A major reason for this inefficiency is the combining of streams for blob and tree packfile items. Combining/chaining streams is like chaining water pipes. The amount of water per unit time that you would be able to transfer through the stream is bottlenecked by the thinnest pipe in the setup. Similarily, the amount of data transferred is bottlenecked by the thinnest stream which in this case was the `GitDeltaManifestEntry` stream.

Streams by design are serial and are meant to yield items one after the other. If you want to absolutely avoid any unnecessary work, this is the best approach. However, this also means severely restricted output. Typically in Mononoke wherever we use streams we poll and prefetch multiple items from the stream while yielding them one after the other so that more concurrent work can be done regardless of the serial nature of the output. The `GitDeltaManifestEntry` stream exists for each commit and contains the metadata of the items that were added or modified as part of this commit. For most commits, this will result in a handful of entries.

So even if a high concurrency of 1000 or 2000 is set, it would not be able to pull those many items concurrently. Instead of using the `GitDeltaManifestEntry` stream as-is, this diff directly collects the stream given that its size is not too large and its safe to hold in memory. With this change, we can potentially hold 6000 `GitDeltaManifests` in memory and then fetch the objects corresponding to those entries with the maximum possible concurrency.

This diff also includes a bunch of "magic" concurrency numbers that I plan to make configurable in the follow up diffs. Note that there are still more optimizations possible but that would require restructuring our data structures so will probably happen after the introduction of `GitHeuristicsManifest` and `GitDeltaManifestV2`

Differential Revision: D55014717

fbshipit-source-id: 14688059eda76195076139b53d0f07a150cf1274