indexedlog: improve lock/mmap compatibility on Windows
Summary:
In the furture we want to create a mmap buffer from the `rlock` lock file.
However, on Windows the locked region can prevent mmap. The `fs2` create
hardcoded the lock range to be `u64::MAX`.

To be compatibile with the upcoming changes, update the lock logic to only lock
the last 1 byte of the u64 range (Windows file lock range can exceed the file
length).

Reviewed By: muirdm

Differential Revision: D54979382

fbshipit-source-id: e131ba809bf336ae021a8d6d463bf172665914db