scsc: add support for --exclude-ancestors-of to scsc log
Summary:
`scsc log` currently supports `descendants-of` via an optional second parameter, however, this makes it hard to implement `exclude-ancestors-of` as this is also an optional parameter.

Instead, we need some way to provide named arguments to things that are already arguments, e.g. `--descendants-of --hg-commit-id HASH` or `--exclude-ancestors-of --bookmark branchname`.

Clap doesn't support this natively, so we emulate this in the same way that we emulate ordering of commit ids in general: using the index of the argument.  We specify named parameters as flag arguments on their own.  If the user positions this argument immediately before a commit specifier, then we name the commit specifier using that flag argument.

The original behaviour of a second positional argument is retained for compatibility, but we now recommend using the named argument.

Differential Revision: D54955128

fbshipit-source-id: 50d05d5b58b516d20c4a536deca2e819c8fd2a5a