Make ClientRequestInfo header optional for MononokeGitServer
Summary: The `MetadataMiddleware` requires that the caller provide the `ClientRequestInfo` header and failing to do so returns in 4xx user error. For MononokeGitServer, the client can be vanilla git binary which does not provide these headers by default. This diff adds a bypass for `MononokeGitServer` so that we can still reuse this middleware.

Reviewed By: markbt

Differential Revision: D55019629

fbshipit-source-id: 9c048db48230b89fef3d1090d1dc741ac848a2bd