indexedlog: rework change detection to avoid race conditions
Summary:
There was an issue with the read reload path of the change detection:

    reload meta (might fail)
    # if someone updates the meta here - change detector will miss the change
    reload change detector u64 (will not fail)

Reordering the reloads helps, but still has issues on error:

    reload change detector u64 (will not fail)
    reload meta (might fail)
    # if reload meta failed, then the change will be missed

Update the change detector u64 to be based on the Log or RotateLog's metadata,
instead of tracking their own. This should avoid the race condition and handle
the error case correctly.

Reviewed By: muirdm

Differential Revision: D55020265

fbshipit-source-id: 1ac06a721a4987f7cae725483d3611eb3be369e3