gitcompat: new crate to talk to git cli
Summary:
Similar to `edenfs-client`, this crate handles talking to Git and various misc
stuff. I think shelling out to `git` is an appealing approach comparing to
libgit2, since things like `git status` can benefit from the latest development
in git upstream.

For now, this crate only has some simple logic to create files needed to treat
the `.git/sl` as a sapling repo. In the future this might include features that
exist in `git.py` like push, pull, and maybe replace the gitstore by
`fast-export`, if we want to drop the libgit2 dependency.

Reviewed By: muirdm

Differential Revision: D54218741

fbshipit-source-id: e7c43474ccf5b7f631179c86b03b451bdc1858bd