git: do not import all references in `.git/sl` mode
Summary:
The `gitdag` used to import all Git references. However, that hurts performance
with too many referecnes. For example, it's a common need to check if a commit
is draft or public, if there are 3000 public heads then the public() calculation
can take a while, which is bad.

By default, `git clone` produces `.git/config` like:

    [remote "origin"]
        url = ...
        fetch = +refs/heads/*:refs/remotes/origin/*

which syncs remote's `refs/heads/*` to local's `refs/remotes/origin/*` on `fetch`.

`sl clone` intentionally does not write this config, and only fetches remote
refs explicitly and does not use wildcards.

So the "too many refs" problem exists for `.git/sl` but not `.sl/git`. This diff
adds logic to ignore refs aggressively. It is a basic version for now. In the
future we might want to be smarter:
- Import names that already exist in metalog `remotenames`.
- Import refs if referred by `HEAD`.
- Do not import local head that has a remote duplicate.
- Look at ref logs (?) if the user explicitly switch branch, import it.

Reviewed By: muirdm

Differential Revision: D54218744

fbshipit-source-id: 811a7abe72ce33cdf65a52f9e4252fe665c63c01