scmstore: fetch file metadata when fetching trees
Summary:
We used to have an option (added in D39833163) to fetch file metadata when fetching trees. This is a big perf win for eden since it avoids subsequent serial remote queries during file system walks. However, it doesn't seem like we ever actually enabled that setting (?). It certainly isn't enabled now (even though the use of it was removed in D46646005).

Anyway, add a new flag scmstore.fetch-tree-metadata which re-enables this functionality. I defaulted it to false since it could also have negative perf consequences for certain commands (it could make the remote queries signifigantly slower).

I also added a config knob eden.fetch-tree-metadata that only applies from the backing store (so we can enable this functionality only for eden, if we want).

Reviewed By: zzl0

Differential Revision: D54976770

fbshipit-source-id: a3db4e8aa35b3122fa826b4ce45a2cabfe425de0