Eden CLI: Suggest to use --debug to get full traceback
Summary:
There's a couple of places in which we just print the exception message without a full trace, making it difficult to troubleshoot where the error is coming from.

In these cases, having the user retry with --debug so they can send us the trace would be useful, so suggest that.

Specifically I want to understand where this is coming from: P1196308487

Reviewed By: kmancini

Differential Revision: D55042054

fbshipit-source-id: 0397e7fe3a2b3d9f0beb022413c52cbc39d6b727