Add a serialize to sort hashes by key
Summary:
D54799402 / D54537233 / D54743888 added a `config_version` column to LFS Server
Scuba logs. That is "working", but not very useful:

{F1470108027}

That's because serializing a hash is non-deterministic, which causes a
different hash to be computed most of the time.

In theory, fixing it should be as simple as decorating `RawRepoConfigs` with
either `preserve_order` (not ideal) or our serializer; but `RawRepoConfigs` is
generated, and the Thrift generator is not *that* customizable AFAICT.

Instead we can wrap the struct with our own and apply the serializer there.
That is arguably safer as it gives us precise control.

Reviewed By: mitrandir77

Differential Revision: D55015775

fbshipit-source-id: 741567f907fe240b98266889736240011b721ea1