debugruntest: improve compatibility with extensions
Summary:
debugruntest executes sl commands in-process to eliminate process startup overhead. This is incompatible with extensions changing during the test since the extensions are only loaded once, installing their function wrappers. If a later command disables the extension, the function wrappers (and other side effects) will still remain, probably messing things up.

In this diff I attempt to fix the function wrappers to be compatible with debugruntest. Basically, when wrapping things we install another layer of wrapping that checks at runtime whether the extension is still enabled (and doesn't run the wrapper if the extension is disabled).

Another approach would be to track all the wrappers and unwrap/rewrap them as we notice extensions have been disbled/enabled, but this way seemed simpler and maybe more reliable.

Reviewed By: quark-zju

Differential Revision: D54839559

fbshipit-source-id: 071bc0e832e48c7049f33c720007ab958a51e933