Unit tests for deletion in submodule expansion
Summary:
## This stack
See D55064943.

## This diff
Adds unit tests for
1. Deleting a submodule in the source repo (repo A),
2. Implicitly deleting **a submodule** in the source repo (repo A)
3. Implicitly deleting directories in the submodule repo (repo B).

The first test is already working as expected, but the other two are not. These will be updated to match the expected behaviour soon, when I work on properly supporting implicit deletes in the submodule expansion.

I'm also adding TODO comments for other test cases I want to add in the next diffs.

Reviewed By: mitrandir77

Differential Revision: D55064942

fbshipit-source-id: e8ea822180337c9ced1b387ecb0a61a164dc6920