derivable_type: rename variants to be consistent in pluralization
Summary:
Derivable types refer to things associated with multiple commits, so typically these are going to be plural.  Either way, we should be consistent to avoid confusion, so let's stick with plurals (or mass nouns in the case of "info").

I have also changed the test manifest types and their string representations.   These should be safe as they're only used in tests.

Reviewed By: RajivTS

Differential Revision: D55065589

fbshipit-source-id: dc2f4232df13905a1d315e4b344710d77bf33b41