expose the pack makeup in tests
Summary:
Let's capture the exact pack composition in the integration test. This way we
can show the changes for better or worse in those integration tests and detect
accidental changes that affect how the packs are made.

From my testing so far the pack generation seems to be deterministic so test
output should be stable.

Reviewed By: RajivTS

Differential Revision: D55122555

fbshipit-source-id: caaf32e60d507f12054d265bf72dde77dc76836f