megarepo: handle "pull -r <xrepo commit>"
Summary:
Support the `rewritepullrev` option for the megarepo cross-repo commit resolution. This incurs an extra remote call even if the "-r" rev belongs to the default remote. We have no way to know a priori if a locally-unknown commit is available in our default remote or not.

An alternative would be to only trigger autopull after the regular pull has failed (that would avoid the extra network request if the regular pull worked). That would be more complicated though, so let's start with this.

Reviewed By: zzl0

Differential Revision: D55045406

fbshipit-source-id: db842a5810eb912c173226d181520eaee0249f96