Fix diff view provider using wrong ctx for cat
Summary:
We were using the ctx value for `cat` commands that was used when initializing the repo. However, if you change your cwd, you may get the "wrong" cwd when running cat, which fails. The right side of the comparison just uses the file URI, which is fine.

instead, we can just use the initial connection ctx for the repo, which we know will be valid for the repo in vscode.

Differential Revision: D55146168

fbshipit-source-id: d9678708540979c4375f5875825f088cac132005