scmstore: add FetchMode to fetch tree children metadata
Summary: Add a new fetch mode that will fetch `children_metadata` from the EdenApi trees endpoint. This metadata gets written to the file aux data cache. This will be used by eden when it is fetching directories that it is confident the per-file aux data will be needed. We don't fetch the children metadata all the time since it is relatively expensive for the server.

Reviewed By: quark-zju

Differential Revision: D55133633

fbshipit-source-id: 5c92edc31bbba439a317db590756994b5aad2f0e