Fix bug triggered by replacing a file with a submodule
Summary:
## This stack
See D55064943.

## This diff
Fixes the 1st issue raised in D55068950.
Other issue (generating deletion in the bonsai) will be fixed in next diff.

## Context on the bug
See docs from `get_previous_submodule_commits`. We need to get the bonsai changeset ids from the submodule commits that the small repo previously depended on to generate the correct delta of file changes.

In this process, we're assuming that any file in the submodule path was of type GitSubmodule and thus contained the git hash representing a submodule pointer.
This isn't always the case, so we should check and treat the submodule as being added if the file in the path is not of type submodule.

Differential Revision: D55124224

fbshipit-source-id: 45761c8636a7a20d0537bb7d6a5442d0fa9d53b6