Derive all data types for large repo in unit tests
Summary:
## This stack
See D55064943.

## This diff

Fixes a gap mentioned in D55068950: making sure that we can derive all the types for all commits in the large repo.
This is to avoid any unexpected issues that can be specific to a single type, like the file being replaced by a directory in D55068950.

This enables the derivation of all types in the config and after syncing a changeset to the large repo and getting all the changeset data, it derives all the types.

**Note:** I also changed some parameter types from `Vec` to slices, so I don't have to keep cloning results throughout the tests.

#thanks pierrec for making it so easy to enable and derive all Mononoke derived data types. Your changes will ensure that any type added in the future will be automatically tested against our x-repo infra!

Differential Revision: D55128490

fbshipit-source-id: 604fa6d017fe28f9e60865060570940f693fd7df