util: use stub hostname in tests
Summary: Tests probably want a stable hostname. But the real reason I'm doing this is that I discovered whoami::hostname() silently crashes on mac, sometimes. I assume it is the same fork detection we have seen in some mac libraries in the past. This may also be the recent reason why chg no longer works on the mac (similar symptom: crashes silently with no output).

Reviewed By: sggutier

Differential Revision: D55171091

fbshipit-source-id: f39980fffdbe226fe0b6e2f022582656dba51bc5