workingcopy: construct DotGitFilesystem
Summary:
This allows the Rust `status` to go through `git status`.

The use of the "eden dirstate" seems to make sense as it's a "disposable",
"overlay" state on top of the "git status" output. We might want to rename
"eden dirstate" to "overlay dirstate" and remove comments about it's
deprecation.

Reviewed By: muirdm

Differential Revision: D54218749

fbshipit-source-id: dd6b37aacbd71a518469904465b3cac822f2f459