treestate: rename legacy_eden_dirstate to overlay_dirstate
Summary:
The "overlay" dirstate is different from the main tree state in a way that it
tracks an "overlay" of changes on top of some other "source of truth" (like
edenfs).

For example, if the underlying "working copy client" reports that "a.txt" is
modified, "b.txt" is new, untracked; then the overlay does not need to track
"a.txt" again, and only records a "b.txt" entry if the user runs "add" to turn
"?" into "A" state.

The "overlay" is meant to be small (only overlay on top of "status") so it can
afford O(N) reloading. It also has extra needs like what to do when the
dirstate p1 is out of sync with the source of truth p1.

I plan to use it for the DotGit support. It's no longer edenfs exclusive so it
gets renamed.

Reviewed By: muirdm

Differential Revision: D55099521

fbshipit-source-id: 6add33741b812bd1279722817e473398bc6d1649