make hg invocations timeout instead of hang
Summary:
A various small set of tests keep timing out on the macOS release job. It's not
always the exact same tests, but some related set of hg integration tests.

The issue does not repro locally. It also does not repro if I run only the
commonly failing tests on a sandcastle job even with stress runs. I have
narrowed the hangs down to `hg add` hanging when setting up the backing repo
for the tests. When we add logging to hg the issue no longer repro the same -
some hg split tests hang but it's unclear if the hangs are still the same issue.

So the hangs seem to be related to some contention that only repros on
sandcastle and is somesort of race condition.

Debugging this fully is going to take some time, the best idea we have is to try
to catch a test while it's hanging and backtrace hg. For now we have disabled
tests on the macOS build so that we can get releases.

Adding a timeout seems to allow all the tests to pass, so we can at least run
some tests on the mac release again while we investigate.

Reviewed By: genevievehelsel

Differential Revision: D55203610

fbshipit-source-id: 07b97c6e3f17013102ee575434c7ab465b7ded02