gitimport: Run derivation concurrently with import
Summary:
I'm currently importing chromium/src, which is a large repo (~2.32M commits) in batches of 1000 commits.

The time it takes is about 11 mins to import a batch of commits followed by 5 mins to finalize that batch, repeating.

At the point we are finalizing a batch (deriving all derived data types for it and writing the mappings), we could start
importing the next batch, but we were not doing that before this diff.

After this diff, we use this opportunity to save ~30% of the time when importing a repo.
Basically, we get data derivation backfilling for free after this change and are back to being IO bound on the time it takes to import all the commits.

Reviewed By: markbt

Differential Revision: D55203295

fbshipit-source-id: 07e1a5311001b8bd8314fec03dc0e97d40de7867