Fix unit tests by making them not dependent on the exact hash
Summary:
I was lazy when introducing these tests and hardcoded the exact hashes at the time.

I was counting on empty field to be skipped, but it turns out that's not the case. It has been a feature request for serde for years, but it hasn't materialized yet.
In theory one could annotate individual fields to tell serde to skip when null, but we can't do that on Thrift-generated files. And one could use serde_with to annotate a struct. But there is no annotation that works on structs that embed other structs.

Instead we will run tests repeatedly and count how many different hashes we get; we expect one.

Reviewed By: gustavoavena

Differential Revision: D55217168

fbshipit-source-id: a9d339bfb12c5fc69cda7bd32aeae04bda67163b