Back out "gitimport: Run derivation concurrently with import"
Summary:
Original commit changeset: 07e1a5311001

Original Phabricator Diff: D55203295

Reason:
I tried it on `chromium/src` and while it went well for a couple of batches of 1000 commits, it eventually started hanging at the end of batches again, and for much longer than before. My guess is that trying to write to the bonsai git mapping dbs while we're also reading from them slows down the process overall as we're racing for the lock. I will backout of this diff for now. To re-introduce it, we will first need to batch the db reads.

Reviewed By: mzr

Differential Revision: D55245809

fbshipit-source-id: 4ff8e849fdcbb9f974a43a3336ce87d1c44ab650