walker: allow specification of changeset keys with known filenodes
Summary:
Currently command arguments that take nodes that include changeset keys assume that filenodes are not known.  This means exclude/include commands cannot specify the situation where filenodes *are* known.

Extend the parser for changeset keys such that a `+` suffix means `filenode_known_derived = true`.  This is a bit of a hack, but we need some way to distinguish this.

Reviewed By: mzr

Differential Revision: D55252757

fbshipit-source-id: 9b354b3cc2ab788aa9efc850a89670805fe1d8ed