Remove commits for focus mode earlier in dag computation
Summary:
Another bug in focus mode: if you try to drag a drop a commit in your focused stack outside your stack, it won't render the preview commit, because the optimistic preview commit is being hidden from the dag rendering subset because of focus mode after the optimistic state is calculated.

This is because we were setting the focus mode handling in the rendering subset, instead of before opstimistic and preview state is applied.

Instead, we move the focus mode computation to `dagWithPreviews`. I think it's reasonable that `dagWithPreviews` excludes focus-mode-ignored commits. It's kind of like a preview, right?

The original dag continues to have all the commits. The mutation dag is also not affected.

This does affect the selectable commit range. So you can no longer use the arrow keys to select "outside" of the focus mode. I think that's actually reasonable, it was  a bit weird that you could before.

Reviewed By: quark-zju

Differential Revision: D55349201

fbshipit-source-id: 53e8dbef18f3d60e3275118bdbfad4f9dd853719