origbackup: auto fix wrong dotdir prefix
Summary:
The `origbackuppath` path is currently statically configured as
`.sl/origbackups` or `.hg/origbackups`. However, when running in a `.git` repo
that triggers a "backup" it will create the `.sl/origbackups` path, which will
trick the next command to use the incomplate and broken `.sl` as the repo
dotdir.

Ideally we can make breaking changes to the `origbackups` config somehow to
avoid this issue. For now I just make it "smarter" by auto correct the dotdir.

Reviewed By: zzl0

Differential Revision: D54218767

fbshipit-source-id: 6a38561d0c28c50f55e4d716188f940c42da8c0f