workingcopy: use git-read-tree to update index after set_parents
Summary:
The `set_parents` API is intended to change the commit used by the "dirstate",
without changing files in the working copy. For Git, the "index" tracks file
contents. Updating the HEAD ref alone is not enough - the index still have
file contents in the old commit.

This diff fixes the issue by using `git-read-tree` to force the `index` to
match the `tree`. Files in the working copy are not changed so if they are
different they will still show up in `status`.

Reviewed By: zzl0

Differential Revision: D54218775

fbshipit-source-id: 02789dfca0873c826c69f0fd19701a6fc0895fbe