namespaces: ignore titles namespaces when conflicting to autopull
Summary:
Sometimes when a user runs `goto releases/release-1.0` and there is a commit
that has a title that matches "releases/release-1.0" they want to trigger
autopull instead.

This diff adds extra checks in the titles namespace logic to avoid conflicts
with autopull. Ideally the namespaces and autopull can be abstracted in a same
layer with different priorities to avoid duplicated caclulation about autopull.
However, that is a larger change that is out of scope of this diff.

Reviewed By: zzl0

Differential Revision: D55436295

fbshipit-source-id: f276b9443c14dbc7b17986febaaab0cb74347fe1