Support fetching stable locations in log
Summary:
We want to support fetching additional commits in `sl log` than just the default set from `smartlog()`. This means providing some api to set which stable locations we care about. Right now, this is just a piece of state in the shared repo you can set (shared because the log calls are shared).

The configured stable locations each include a hash, which is what gets added into the log revest. This ensures ISL shows the real commit.

While doing this, I refactored the commit fetching revset string manipulation stuff a bit, which should make it easier to ensure we have a well-formed revset string.

Alternate ways we could think about supporting this:
- have the stable location hash be a "fake" commit, not actually included in the `log` call revset. We'd just fill in some data as a CommitInfo, and re-inject it into the list of commits after the log call goes through. This seems a little hacky to me, though it avoids weirdness if the stable location hash is not known in the repo which can break the entire log call.

Reviewed By: quark-zju

Differential Revision: D55342078

fbshipit-source-id: 5ea26949e2c4492d2ab94b6689f3168b35d68c16